+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Schedule Functions - make experiment-related schedule functions available in `quantify_scheduler.schedules` (!572)
 - Tests - Removed unused `lo0` and added `ttl_acq_threshold` in `qblox_test_mapping_nv_centers.json` so that `TestNVDarkESRSched` suite passes (!579)
 - Documentation - Removed deprecated code from the Compiling to Hardware Tutorial (!582)
+- Qblox backend - Typecast attenuations to `int`s before assigning them (!570)
 
 ## 0.10.1 (2022-12-20)